refactor(users): split /admin/users into workspace + /platform/users (#PR4)
Closes the audit finding from #2305's multi-agent review. The page used
to live at `/admin/users` and branch on `isPlatformAdmin` at ~10 sites:
workspace admins saw "Remove from workspace"; platform admins saw "Ban /
Unban / list all tenants" through the same URL. Role-driven divergence
under one URL was exactly the pattern that triggered the `/platform/*`
top-level lift for the rest of the platform-admin surface.

Pattern matches what #2305 did for organizations + abuse:

  - Extract the page body to `/admin/users/_users-page.tsx` (Next.js
    underscore prefix = not a route). Accepts `scope: "workspace" |
    "platform"` instead of deriving from the caller's role.
  - `/admin/users/page.tsx` becomes a thin wrapper that renders
    `<UsersPage scope="workspace" />`. Platform admins landing here get
    redirected to `/platform/users` so the URL matches the data they're
    operating on.
  - `/platform/users/page.tsx` adds `usePlatformAdminGuard` and renders
    `<UsersPage scope="platform" />`.
  - Nav: new entry "Users (All Tenants)" under the Platform group.

Driving scope from the URL (not the role) means the same component
renders deterministically: a platform_admin standing on `/admin/users`
sees workspace data; a workspace admin trying to visit `/platform/users`
gets bounced by the guard. No more "same URL, different page depending
on who's looking."

Tests in `users-pages-split.test.tsx` pin the four routing invariants:
  - workspace admin on /admin/users → no redirect, renders
  - platform admin on /admin/users → router.replace("/platform/users")
  - /platform/users blocked by guard → renders nothing
  - /platform/users passes guard → renders the page
